Convert api4 helper functionality to a trait & make available
