+++ ext/afform/core/tests/phpunit/Civi/Afform/AfformContactSummaryTest.php
@@ -1,16 +1,19 @@
 <?php
 namespace Civi\Afform;
 
-// Hopefully temporry workaround for loading core test classes
-require_once __DIR__ . '/../../../../../../../tests/phpunit/api/v4/Api4TestBase.php';
-
 use Civi\Api4\Afform;
+use Civi\Test;
+use Civi\Test\Api4TestTrait;
 use Civi\Test\CiviEnvBuilder;
+use Civi\Test\HeadlessInterface;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @group headless
  */
-class AfformContactSummaryTest extends \api\v4\Api4TestBase {
+class AfformContactSummaryTest extends TestCase implements HeadlessInterface {
+
+  use Api4TestTrait;
 
   private $formNames = [
     'contact_summary_test1',
@@ -21,12 +24,12 @@ class AfformContactSummaryTest extends \api\v4\Api4TestBase {
   ];
 
   public function setUpHeadless(): CiviEnvBuilder {
-    return \Civi\Test::headless()->installMe(__DIR__)->install('org.civicrm.search_kit')->apply();
+    return Test::headless()->installMe(__DIR__)->install('org.civicrm.search_kit')->apply();
   }
 
   public function tearDown(): void {
     Afform::revert(FALSE)->addWhere('name', 'IN', $this->formNames)->execute();
-    parent::tearDown();
+    $this->deleteTestRecords();
   }
 
   public function testAfformContactSummaryTab(): void {
